requirements.txt
amqp==5.2.0
asgiref==3.7.2
astroid==3.0.2
asttokens==2.4.1
CacheControl==0.13.1
cachetools==5.3.2
celery==5.3.6
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
CMRESHandler==1.0.0
coreapi==2.3.3
coreschema==0.0.4
cron_descriptor==1.4.0
cryptography==41.0.7
decorator==5.1.1
dill==0.3.7
Django==4.2.9
django-celery-beat==2.5.0
django-cors-headers==4.3.1
django-rest-swagger==2.2.0
django-storages==1.14.2
django-timezone-field==6.1.0
django-utils-six==2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
executing==2.0.1
filelock==3.13.1
firebase-admin==6.3.0
gevent==23.9.1
google-api-core==2.15.0
google-api-python-client==2.113.0
google-auth==2.26.2
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.14.0
google-cloud-storage==2.14.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.62.0
greenlet==3.0.3
grpcio==1.60.0
grpcio-status==1.60.0
httplib2==0.22.0
idna==3.6
importlib-metadata==7.0.1
inflection==0.5.1
ipython==8.26.0
isort==5.13.2
itypes==1.2.0
jmespath==1.0.1
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pycryptodome==3.20.0
Pygments==2.18.0
PyJWT==2.8.0
pylint==3.0.3
pyotp==2.9.0
pyparsing==3.1.1
python-crontab==3.0.0
python-dateutil==2.8.2
python-decouple==3.8
python-http-client==3.3.7
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
rsa==4.9
sqlparse==0.4.4
stack-data==0.6.3
starkbank-ecdsa==2.2.0
tomlkit==0.12.3
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2023.4
uritemplate==4.1.1
urllib3==1.26.18
uWSGI==2.0.23
validators==0.22.0
vine==5.1.0
wcwidth==0.2.13
zipp==3.17.0
zope.event==5.0
zope.interface==6.1
